+++ test/testDeckManager.js
@@ -6,7 +6,8 @@ const sinon = require('sinon');
 const sc = require('sinon-chai');
 chai.use(sc);
 
-const deckManager = rewire('../lib/deckManager');
+// Use require because rewire breaks sinon timers if first
+const deckManager = require('../lib/deckManager');
 
 const PIXEL = "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVQYV2NYlV/1HwAF7QKTgvPu3QAAAABJRU5ErkJggg==";
 
@@ -15,7 +16,7 @@ describe('Deck Manager', function() {
 	beforeEach(function() {
 		deck = {clearAllKeys: sinon.fake(), clearKey: sinon.fake(), ICON_SIZE: 32, KEY_COLUMNS: 2, KEY_ROWS: 2, fillPanel: sinon.fake(), setBrightness: sinon.fake()};
 		buttons = new Array(6).fill(0).map(() => {
-			return {stop: sinon.spy(), start: sinon.spy()};
+			return {stop: sinon.spy(), start: sinon.spy(), activate: sinon.spy()};
 		});
 		// Set one to null
 		buttons[2] = null;
@@ -45,21 +46,38 @@ describe('Deck Manager', function() {
 		});
 	});
 
+	describe('Standard buttons', function() {
+		it('should handle standard buttons', function() {
+			let dm = new deckManager(deck, buttons, {});
+			dm.buttonPressed(0);
+			dm.buttonPressed(2);
+			expect(buttons[2]).to.be.null;
+			expect(buttons[0].activate).to.be.calledOnce;
+		});
+	});
+
 	describe('Screensaver', function() {
-		let dm;
+		let dm, clock;
 		beforeEach(function() {
-			dm = new deckManager(deck, buttons, {screensaver: {animation: require('path').join(__dirname, '../test_resources/blink.gif'), brightness: 10}});
+			clock = sinon.useFakeTimers();
+			dm = new deckManager(deck, buttons, {screensaver: {animation: require('path').join(__dirname, '../test_resources/blink.gif'), brightness: 10, timeoutMinutes: 1}});
+			return dm.screensaver.isReady;
 		});
 		afterEach(function() {
 			dm.stopScreensaver();
+			clock.restore();
+			sinon.restore();
 		});
 		it('should stop all buttons when screensaver started', function() {
-			return dm.screensaver.isReady.then(() => {
-				dm.startScreensaver();
-				buttons.forEach((b) => {
-					if (b) expect(b.stop).to.be.calledOnce;
-				});
+			dm.startScreensaver();
+			buttons.forEach((b) => {
+				if (b) expect(b.stop).to.be.calledOnce;
 			});
+		});
+		it('should trigger a start after a timeout', async function() {
+			expect(dm.ssActive).to.be.false;
+			await clock.tickAsync(61 * 1000);
+			expect(dm.ssActive).to.be.true;
 		});
 		it('should start all buttons when screensaver stopped', function() {
 			dm.stopScreensaver();
@@ -68,22 +86,32 @@ describe('Deck Manager', function() {
 			});
 		});
 		it('should set brightness when screensaver starts', function() {
-			return dm.screensaver.isReady.then(() => {
-				dm.startScreensaver();
-				expect(deck.setBrightness).to.be.calledWith(10);
-			});
+			dm.startScreensaver();
+			expect(deck.setBrightness).to.be.calledWith(10);
 		});
 		it('should not touch brightness if not supplied', function() {
 			delete dm.config.screensaver.brightness;
-			return dm.screensaver.isReady.then(() => {
-				dm.startScreensaver();
-				expect(deck.setBrightness).to.not.be.called;
-			});
+			dm.startScreensaver();
+			expect(deck.setBrightness).to.not.be.called;
 		});
 		it('should restore brightness when screensaver ends', function() {
 			dm.storedBrightness = 45;
 			dm.stopScreensaver();
 			expect(deck.setBrightness).to.be.calledWith(45);
+		});
+		it('should stop a screensaver on button press', function() {
+			dm.startScreensaver();
+			expect(dm.ssActive).to.be.true;
+			dm.buttonPressed(0);
+			expect(dm.ssActive).to.be.false;
+		});
+		it('should stop a screensaver and activate on a second press', function() {
+			dm.startScreensaver();
+			expect(dm.ssActive).to.be.true;
+			dm.buttonPressed(0);
+			expect(dm.ssActive).to.be.false;
+			dm.buttonPressed(0);
+			expect(buttons[0].activate).to.be.calledOnce;
 		});
 	});
 
@@ -131,7 +159,7 @@ describe('Deck Manager', function() {
 			expect(buttons[0]).to.be.null;
 		});
 
-		it('should preserve sticky buttons', function(){
+		it('should preserve sticky buttons', function() {
 			dm = new deckManager(deck, buttons, {
 				sticky: [{keyIndex: 0, icon: PIXEL, command: 'none'}],
 				pages: [{
@@ -146,7 +174,11 @@ describe('Deck Manager', function() {
 	});
 
 	describe('Dynamic page', function() {
-		let dm, spawn, revert, so;
+		let dm, spawn, revert, so, deckManager;
+		// Do rewire
+		before(function() {
+			deckManager = rewire('../lib/deckManager');
+		});
 		beforeEach(function() {
 			so = {
 				stdout: new events.EventEmitter()
@@ -177,7 +209,7 @@ describe('Deck Manager', function() {
 		it('should load a page based on command output', function() {
 			dm.changePage('testPage');
 			buttons.forEach((b) => {
-				if ( b ) {
+				if (b) {
 					expect(b.btnCfg.keyIndex).to.eql(1);
 				} else {
 					expect(b).to.be.null;
@@ -207,7 +239,7 @@ describe('Deck Manager', function() {
 			let con = sinon.stub(console, 'error');
 			dm.changePage('testPage');
 			buttons.forEach((b) => {
-				if ( b ) {
+				if (b) {
 					expect(b.btnCfg.keyIndex).to.eql(1);
 				} else {
 					expect(b).to.be.null;
@@ -216,7 +248,7 @@ describe('Deck Manager', function() {
 			expect(con).to.not.be.called;
 			so.stdout.emit('data', 'garbage');
 			buttons.forEach((b) => {
-				if ( b ) {
+				if (b) {
 					expect(b.btnCfg.keyIndex).to.eql(1);
 				} else {
 					expect(b).to.be.null;